+++ examples/hello_world.rs
@@ -2,7 +2,7 @@ use aws_lambda_runtime_proxy::Proxy;
 
 #[tokio::main]
 async fn main() {
-  let proxy = Proxy::default().spawn().await;
+  let proxy = Proxy::default().spawn().await.unwrap();
   // equals to:
   // let proxy = Proxy::default()
   //   .port(3000)