+++ src/client.rs
@@ -1,3 +1,4 @@
+use anyhow::Result;
 use http_body_util::{BodyExt, Full};
 use hyper::{
   body::{Body, Bytes, Incoming},
@@ -12,6 +13,7 @@ use tracing::error;
 /// An http client for the Lambda Runtime API.
 pub struct LambdaRuntimeApiClient<ReqBody>(SendRequest<ReqBody>);
 
+// TODO: use AsRef instead of Deref?
 impl<ReqBody> Deref for LambdaRuntimeApiClient<ReqBody> {
   type Target = SendRequest<ReqBody>;
   fn deref(&self) -> &Self::Target {
@@ -26,45 +28,40 @@ impl<ReqBody> DerefMut for LambdaRuntimeApiClient<ReqBody> {
 
 impl<ReqBody: Body + Send + 'static> LambdaRuntimeApiClient<ReqBody> {
   /// Create a new client and connect to the runtime API.
-  pub async fn new() -> Self
+  pub async fn new() -> Result<Self>
   where
     ReqBody::Data: Send,
     ReqBody::Error: Into<Box<dyn std::error::Error + Send + Sync>>,
   {
-    let address =
-      std::env::var("AWS_LAMBDA_RUNTIME_API").expect("Missing AWS_LAMBDA_RUNTIME_API env var");
+    let address = std::env::var("AWS_LAMBDA_RUNTIME_API")?; // TODO: cache the result?
 
-    let stream = TcpStream::connect(address)
-      .await
-      .expect("Failed to connect to runtime API");
+    // TODO: re-use the connection?
+    let stream = TcpStream::connect(address).await?;
     let io = TokioIo::new(stream);
-    let (sender, conn) = http1::handshake(io)
-      .await
-      .expect("Failed to handshake with runtime API");
+    let (sender, conn) = http1::handshake(io).await?;
 
     // Spawn a task to poll the connection, driving the HTTP state
-    tokio::task::spawn(async move {
+    tokio::spawn(async move {
       if let Err(err) = conn.await {
         error!("Connection failed: {:?}", err);
       }
     });
 
-    Self(sender)
+    Ok(Self(sender))
   }
 }
 
 impl LambdaRuntimeApiClient<Incoming> {
   /// Send a request to the runtime API and return the response.
-  pub async fn forward(req: Request<Incoming>) -> hyper::Result<Response<Full<Bytes>>> {
+  pub async fn forward(req: Request<Incoming>) -> Result<Response<Full<Bytes>>> {
     // tested and it looks like we create the client every time is faster than lock a Arc<Mutex<>> and reuse it.
     // create a new client and send the request usually cost < 1ms.
     let res = LambdaRuntimeApiClient::new()
-      .await
+      .await?
       .send_request(req)
-      .await
-      .unwrap();
+      .await?;
     let (parts, body) = res.into_parts();
-    let bytes = body.collect().await.unwrap().to_bytes();
+    let bytes = body.collect().await?.to_bytes();
     Ok(Response::from_parts(parts, Full::new(bytes)))
 
     // TODO: why we can't just return `LambdaRuntimeApiClient::new().await.send_request(req).await`?