+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+- **_Breaking Change_**: apply `anyhow` for better error handling.
 - Feat: apply `tracing` instead of just `println` on error.
 
 ## v0.2.1